Refactor test helper to testing file
